refactor(bridge): move text payload ownership
