+++ whatsrust/lib/main.go
@@ -17,10 +17,6 @@ typedef struct {
 } Contact;
 
 typedef struct {
-	char* text;
-} TextMessage;
-
-typedef struct {
 	uint8_t kind;
 	char* path;
 	char* fileID;
@@ -84,8 +80,6 @@ static void callHistorySync(HistorySyncHandler hdl, uint32_t percent) {
 */
 import "C"
 import (
-	"unsafe"
-
 	"go.mau.fi/whatsmeow/proto/waE2E"
 	"go.mau.fi/whatsmeow/types"
 )
@@ -122,22 +116,6 @@ const (
 	FileTypeDocument
 	FileTypeSticker
 )
-
-func emitTextMessage(cinfo C.MessageInfo, text string, isSync bool) {
-	ctext := C.CString(text)
-	defer C.free(unsafe.Pointer(ctext))
-
-	content := (*C.TextMessage)(C.malloc(C.sizeof_TextMessage))
-	content.text = ctext
-	defer C.free(unsafe.Pointer(content))
-
-	message := C.Message{
-		info:        cinfo,
-		messageType: C.uint8_t(MessageTypeText),
-		message:     unsafe.Pointer(content),
-	}
-	C.callMessageHandler(messageHandler, C.bool(isSync), &message)
-}
 
 func HandleMessage(info types.MessageInfo, msg *waE2E.Message, isSync bool) {
 	msg, viewOnceUnavailable := unavailableViewOnceMessage(msg)
